refactor size representation
This commit changes the `Size` type
to represent an amount of bytes instead
of an amount of bits.

The main reason for this change are buffer
allocations. It is expected that a user
can allocate a buffer with a pre-defined size/cap
easily via:
```
buffer := make([]byte, 64 * mem.KB)
```

When representing `Size` as number of bits, a user
would allocate an `8 * 1000` byte buffer - since a
`KB` would be 8000.
This is pitfall a user could easily fall into.

Since most Go code deals with bytes anyways we
re-define `Size` to represent the number of bytes.

For measuring sizes in bits we could add a new type,
e.g. `BitSize`, anytime.

As part of this change all formatting / parsing of sizes
does no longer support de/serialization to/from bits.
For example, "1Mbit" is no longer a valid size.

Signed-off-by: Andreas Auernhammer <hi@aead.dev>